Optimized code for controller, update on test result, add search and sort for each entity

+++ hiv-treatment-support-system/backend/src/main/java/backend/schedule/model/Schedule.java
@@ -26,10 +26,14 @@ import lombok.NoArgsConstructor;
 public class Schedule {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
+
     private String type;
+
     private String status;
+
     private LocalDate date;
+    
     private LocalTime slot;
 
     @OneToOne
